make typescript an optional dependency, if it fails, the plugin does nothing
