+++ package.json
@@ -2,21 +2,17 @@
   "name": "@joshwooding/vite-plugin-react-docgen-typescript",
   "version": "0.1.0",
   "description": "A vite plugin to inject react typescript docgen information",
-  "author": "Josh Wooding",
-  "license": "MIT",
+  "keywords": [
+    "vite",
+    "vite-plugin",
+    "react-docgen-typescript"
+  ],
   "repository": {
     "type": "git",
     "url": "https://github.com/joshwooding/vite-plugin-react-docgen-typescript.git"
   },
-  "files": [
-    "dist/**/*",
-    "README.md",
-    "*.js",
-    "*.d.ts"
-  ],
-  "main": "./dist/index.cjs",
-  "module": "./dist/index.mjs",
-  "types": "./dist/index.d.ts",
+  "license": "MIT",
+  "author": "Josh Wooding",
   "exports": {
     ".": {
       "types": "./dist/index.d.ts",
@@ -24,15 +20,20 @@
       "require": "./dist/index.cjs"
     }
   },
+  "main": "./dist/index.cjs",
+  "module": "./dist/index.mjs",
+  "types": "./dist/index.d.ts",
+  "files": [
+    "dist/**/*",
+    "README.md",
+    "*.js",
+    "*.d.ts"
+  ],
   "scripts": {
     "build": "unbuild",
     "rome": "rome format . --write",
     "rome:check": "rome format .",
     "test": "vitest"
-  },
-  "peerDependencies": {
-    "typescript": ">= 4.3.x",
-    "vite": "^3.0.0"
   },
   "dependencies": {
     "glob": "^7.2.0",
@@ -50,10 +51,14 @@
     "vite": "^3.2.5",
     "vitest": "^0.25.8"
   },
-  "keywords": [
-    "vite",
-    "vite-plugin",
-    "react-docgen-typescript"
-  ],
+  "peerDependencies": {
+    "typescript": ">= 4.3.x",
+    "vite": "^3.0.0"
+  },
+  "peerDependenciesMeta": {
+    "typescript": {
+      "optional": true
+    }
+  },
   "packageManager": "yarn@3.2.4"
 }
