+++ src/utils/filter.ts
@@ -1,0 +1,5 @@
+import { PropFilter } from "react-docgen-typescript/lib/parser";
+
+export const defaultPropFilter: PropFilter = (prop) => {
+	return !prop.parent?.fileName.includes("node_modules");
+};
